% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ama_csl.R
\name{add_ama_csl}
\alias{add_ama_csl}
\title{Add file csl AMA style}
\usage{
add_ama_csl(folder_name)
}
\arguments{
\item{folder_name}{Way to file}
}
\value{
Append file hlper.R
}
\description{
Add file csl AMA style
}
