% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_article_frame.R
\name{add_article_frame}
\alias{add_article_frame}
\title{Append article frame to file}
\usage{
add_article_frame(file_name)
}
\arguments{
\item{file_name}{Way to file}
}
\value{
Append article frame to file. Mostly for .qmd files
}
\description{
Append article frame to file
}
\examples{
add_article_frame("./article.qmd")
}
