% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gost_csl.R
\name{add_gost_csl}
\alias{add_gost_csl}
\title{Add file csl GOST style}
\usage{
add_gost_csl(folder_name)
}
\arguments{
\item{folder_name}{Way to file}
}
\value{
Append file hlper.R
}
\description{
Add file csl GOST style
}
