% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_helper.R
\name{add_helper}
\alias{add_helper}
\title{Add file with hel srtucture}
\usage{
add_helper(folder_name)
}
\arguments{
\item{folder_name}{Way to file}
}
\value{
Append file hlper.R
}
\description{
Add file with hel srtucture
}
