% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lec_protocol_frame.R
\name{add_lec_protocol_frame}
\alias{add_lec_protocol_frame}
\title{Append lec_protocol frame to file}
\usage{
add_lec_protocol_frame(file_name)
}
\arguments{
\item{file_name}{Way to file}
}
\value{
Append lec_protocol frame to file. Mostly for .qmd files
}
\description{
Append lec_protocol frame to file
}
\examples{
add_lec_protocol_frame("./inst/extdata/lec_protocol.qmd")
}
