% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sc_protocol_frame.R
\name{add_sc_protocol_frame}
\alias{add_sc_protocol_frame}
\title{Append sc_protocol frame to file}
\usage{
add_sc_protocol_frame(file_name)
}
\arguments{
\item{file_name}{Way to file}
}
\value{
Append sc_protocol frame to file. Mostly for .qmd files
}
\description{
Append sc_protocol frame to file
}
