% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tesis_frame.R
\name{add_tesis_frame}
\alias{add_tesis_frame}
\title{Append tesis frame to file}
\usage{
add_tesis_frame(file_name)
}
\arguments{
\item{file_name}{Way to file}
}
\value{
Append tesis frame to file. Mostly for .qmd files
}
\description{
Append tesis frame to file
}
\examples{
add_tesis_frame("./inst/extdata/tesis.qmd")
}
