% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_calc_block.R
\name{create_calc_block}
\alias{create_calc_block}
\title{Create folders and frame to calculate}
\usage{
create_calc_block(path_folder)
}
\arguments{
\item{path_folder}{Way to folder where create block}
}
\value{
Folders to calculate
}
\description{
Create folders and frame to calculate
}
