% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_disser_frame.R
\name{create_disser_frame}
\alias{create_disser_frame}
\title{Create disser frame by heading}
\usage{
create_disser_frame(file_name)
}
\arguments{
\item{file_name}{Adress to json file}
}
\value{
Create .qmd files to disser
}
\description{
Create disser frame by heading
}
\examples{
folder <- "./inst/extdata/"
file_name <- "./inst/extdata/plan.json"
create_disser_frame(folder,file_name)
}
