% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_disser_str.R
\name{create_disser_str}
\alias{create_disser_str}
\title{Create config file}
\usage{
create_disser_str(folder)
}
\arguments{
\item{folder}{way to folder}
}
\value{
Create json for create_disser_frame
}
\description{
Create config file
}
