% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_num.R
\name{create_table_num}
\alias{create_table_num}
\title{Create df with numeric measures}
\usage{
create_table_num(df, dev, transl, normuse = TRUE)
}
\arguments{
\item{df}{Exploring dataframe}

\item{dev}{Vector with devide}

\item{transl}{Vector with names to final text}

\item{normuse}{Boolean, for using normality or not.}
}
\value{
df with numeric measures
}
\description{
Create df with numeric measures
}
\examples{
# df <- dplyr::tibble(
# a = as.factor(c("D","F","F","D")),
# b = c(1:4),
# c = c(5:8)
# dev <- "a"
# transl <- с("group","gender","head color")
# create_table_num(df, dev, transl)
}
