% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_csv_param.R
\name{dev_csv_param}
\alias{dev_csv_param}
\title{Split CSV by group into separate files}
\usage{
dev_csv_param(file_in, folder, dev = "параметр")
}
\arguments{
\item{file_in}{Path to input CSV file}

\item{folder}{Output folder for split files}

\item{dev}{Name of the column to split by (as string)}
}
\value{
Invisibly returns the input data frame
}
\description{
Split CSV by group into separate files
}
\examples{
\dontrun{
dev_csv_param("data.csv", "split", dev = "параметр")
}
}
