% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escribir.R
\name{escribir}
\alias{escribir}
\title{Type in file string with new line break}
\usage{
escribir(text, to_file)
}
\arguments{
\item{text}{A string of text}

\item{to_file}{Way to file}
}
\value{
Append string in file
}
\description{
Type in file string with new line break
}
\examples{
# escribir("Hello world!", "./inst/extdata/textfile.txt")
}
