% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escribir_de_file.R
\name{escribir_de_file}
\alias{escribir_de_file}
\title{Type all text from file}
\usage{
escribir_de_file(way_to_text, to_file)
}
\arguments{
\item{way_to_text}{Way to text}

\item{to_file}{Way to file}
}
\value{
Append text from file to target file
}
\description{
Type all text from file
}
