% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_key.R
\name{find_key}
\alias{find_key}
\title{Rewrite bib keys}
\usage{
find_key(link_file, bib_file, n)
}
\arguments{
\item{link_file}{Links from file}

\item{bib_file}{Actual bib File}

\item{n}{Number in text}
}
\value{
String for table
}
\description{
Rewrite bib keys
}
\examples{
# links <- "links.xlsx"
# bib_lnks <- "references.bib"
# n <- 34
}
