% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_disserwork.R
\name{guide_disserwork}
\alias{guide_disserwork}
\title{Create disser frame by heading}
\usage{
guide_disserwork()
}
\value{
Type instruction to use this package
}
\description{
Create disser frame by heading
}
