% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how_many.R
\name{how_many}
\alias{how_many}
\title{Type all text from file}
\usage{
how_many(vector, descripror, file_name)
}
\arguments{
\item{vector}{Vector for calculation}

\item{descripror}{Introductory explanation for forming relevant text}

\item{file_name}{Introductory explanation for forming relevant text}
}
\value{
String with all cases to target file
}
\description{
Type all text from file
}
\examples{
# df <- mtcars
# file.create("./text.txt")
# file_name = "./text.txt"
# how_many(df$mpg,"Всего в исследование вошло",file_name)
}
