% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_col_names.R
\name{print_col_names}
\alias{print_col_names}
\title{Print colnames}
\usage{
print_col_names(df)
}
\arguments{
\item{df}{Data frame from which needs colnames}
}
\value{
formated string in terminal
}
\description{
Print colnames
}
\examples{
df <- mtcars
print_col_names(df)
}
