% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_text_cat.R
\name{print_text_cat}
\alias{print_text_cat}
\title{Create text from df with categorical results}
\usage{
print_text_cat(df, filename1)
}
\arguments{
\item{df}{df with categorical results}

\item{filename1}{way to save text in file}
}
\value{
text in file with text
}
\description{
Create text from df with categorical results
}
