% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_text_num.R
\name{print_text_num}
\alias{print_text_num}
\title{Create text from df with numeral results}
\usage{
print_text_num(df, filename1, mes)
}
\arguments{
\item{df}{df with numeral results}

\item{filename1}{way to save text in file}

\item{mes}{vector with names of measures}
}
\value{
text in file with text
}
\description{
Create text from df with numeral results
}
