% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_discr.R
\name{quantity_discr}
\alias{quantity_discr}
\title{Type categoric data data to string}
\usage{
quantity_discr(parametr, vector, file_name)
}
\arguments{
\item{parametr}{Which parameter}

\item{vector}{Vector for calculation}

\item{file_name}{Introductory explanation for forming relevant text}
}
\value{
Print categoric data data to string
}
\description{
Type categoric data data to string
}
\examples{
parametr = "Engine is "
file_name <- "text.txt"
quantity_discr(parametr = parametr, vector = as.factor(mtcars$vs),file_name)
}
