% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_discr_short.R
\name{quantity_discr_short}
\alias{quantity_discr_short}
\title{Type categoric data in short form data to string}
\usage{
quantity_discr_short(parametr, vector, file_name)
}
\arguments{
\item{parametr}{Which parameter}

\item{vector}{Vector for calculation}

\item{file_name}{Introductory explanation for forming relevant text}
}
\value{
Print categoric data in short form data to string
}
\description{
Type categoric data in short form data to string
}
\examples{
# parametr = "Engine is "
# file_name <- "text.txt"
# quantity_discr_short(parametr = parametr, vector = as.factor(mtcars$vs),file_name)
}
