% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_check.R
\name{review_check}
\alias{review_check}
\title{Check review chapters}
\usage{
review_check(pages, actual_year, file_Rmd, file_bib)
}
\arguments{
\item{pages}{Target pages}

\item{actual_year}{Actual year poblications}

\item{file_Rmd}{To target file Rmd}

\item{file_bib}{To target file bib}
}
\value{
String for table
}
\description{
Check review chapters
}
\examples{
# pages <- 200 # целевое значение страниц
# actual_year <- 2015 # уровень актуальности
# file_Rmd <- "archive/Text_work/Chapters/full_text/Abus_full.Rmd"
# file_bib <- "archive/Text_work/Chapters/full_text/references.bib"

# review_check(pages,actual_year,file_Rmd,file_bib)
}
