% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_num_norm.R
\name{row_num_norm}
\alias{row_num_norm}
\title{Type norm data table row}
\usage{
row_num_norm(parametr, vector, file_name)
}
\arguments{
\item{parametr}{Which parameter}

\item{vector}{vector Vector for calculation}

\item{file_name}{To target file}
}
\value{
String for table
}
\description{
Type norm data table row
}
\examples{
# df <- mtcars
# file.create("./text.txt")
# file_name = "./text.txt"
# row_num_norm("Miles/(US) gallon ",df$mpg,file_name)
}
