% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_num_nonorm.R
\name{str_num_nonorm}
\alias{str_num_nonorm}
\title{Type no norm data string}
\usage{
str_num_nonorm(vector, parametr, descripror, mesurement, file_name)
}
\arguments{
\item{vector}{Vector for calculation}

\item{parametr}{Which parameter}

\item{descripror}{Introductory explanation for forming relevant text}

\item{mesurement}{Unit of measurement}

\item{file_name}{Introductory explanation for forming relevant text}
}
\value{
Append text from file to target file
}
\description{
Type no norm data string
}
\examples{
df <- mtcars
file.create("text.txt")
file_name = "text.txt"
str_num_nonorm(df$mpg, "Miles/(US) gallon ",
             "во всей выборке составил ","miles",file_name)
}
