% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_fact.R
\name{create_table_fact}
\alias{create_table_fact}
\title{Create df with factor measures}
\usage{
create_table_fact(df, dev, transl)
}
\arguments{
\item{df}{Exploring dataframe}

\item{dev}{Vector with devide}

\item{transl}{Vector with names to final text}
}
\value{
df with factor measures
}
\description{
Create df with factor measures
}
\examples{
df <- tibble(
a = as.factor("D","F","F","D"),
b = as.factor("male","female","female","male"),
c = as.factor("blond","redhead","blomd","brunet"))
dev <- df$a
transl <- Ñ("group","gender","head color")
create_table_fact(df, dev, transl)
}
