% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_fact.R
\name{create_table_fact}
\alias{create_table_fact}
\title{Create df with factor measures}
\usage{
create_table_fact(df, dev, transl)
}
\arguments{
\item{df}{Exploring dataframe}

\item{dev}{Vector with devide}

\item{transl}{Vector with names to final text}
}
\value{
df with factor measures
}
\description{
Create df with factor measures
}
\examples{
df <- dplyr::tibble(
a = as.factor(c("D","F","F","D")),
b = as.factor(c("male","female","female","male")),
c = as.factor(c("blond","redhead","blomd","brunet")))
dev <- "a"
transl <- c("group","gender","head color")
create_table_fact(df, dev, transl)
}
